+++ flutter/lib/src/native/platform/functions_registerer.dart
@@ -5,6 +5,7 @@ import 'constraints.dart' as constraints;
 import 'ice_candidate.dart' as ice_candidate;
 import 'ice_candidate_error.dart' as ice_candidate_error;
 import 'ice_servers.dart' as ice_servers;
+import 'logging.dart' as logging;
 import 'media_device_info.dart' as media_device_info;
 import 'media_devices.dart' as media_devices;
 import 'media_display_info.dart' as media_display_info;
@@ -19,6 +20,7 @@ import 'transport.dart' as transport;
 /// Registers functions needed for platform utils working.
 void registerFunctions(DynamicLibrary dl) {
   object.registerFunctions(dl);
+  logging.registerFunctions(dl);
   media_track.registerFunctions(dl);
   peer_connection.registerFunctions(dl);
   transceiver.registerFunctions(dl);