+++ CHANGELOG.md
@@ -10,6 +10,10 @@ All user visible changes to this project will be documented in this file. This p
 
 [Diff](https://github.com/instrumentisto/medea-jason/compare/medea-jason-0.12.0...master)
 
+### Added
+
+- `LogLevel` enum and `set_log_level(LogLevel)` global function to configure global library logging level. ([#260])
+
 ### Changed
 
 - Disabled audio processing in captured system audio on web (already disabled on desktops). ([#259])
@@ -21,6 +25,7 @@ All user visible changes to this project will be documented in this file. This p
 
 [#258]: https://github.com/instrumentisto/medea-jason/pull/258
 [#259]: https://github.com/instrumentisto/medea-jason/pull/259
+[#260]: https://github.com/instrumentisto/medea-jason/pull/260